Apply comments: reuse shape from traits
